Updated plugin to fix a UI glitch, plus linted the code
The UI glitch was that the initial screen (`LoadRWallet`) would have a
possible "stale" view of the coins in the wallet if the wallet was up
for a while with the plugin loaded, and it got sent new coins in the
meantime, so the transfer "tx/hr" stat would be inaccurate.

So, in this commit, the initial screen was updated to listen for the
`history_updated_signal` and to update `LoadRWallet`'s self.utxos when
that signal arrives, so that the stats in the UI would be accurate.

Also in this commit: Did some python code linting to get the code more
in-line with pep8.