Disable multisig wallets
The XPUB for multisig wallets doesn't work anyway for generating
P2PKH addresses -- if we were to use it the funds would be dangerously sent to
hard-to-get-to P2PKH addresses, requiring advanced support to recover.

Also as a transfer source multisig doesn't work.  So I added code to
detect multisig and to prevent them from being used.